+++ src/settings.js
@@ -431,7 +431,7 @@ const SETTINGS = {
         "no-item-macros": {
           name: "ddb-importer.settings.no-item-macros.name",
           hint: "ddb-importer.settings.no-item-macros.hint",
-          config: true,
+          config: false,
           type: Boolean,
           default: false,
         },